refactor(app): useAppServerEvents 一期拆四族
- 3843→1981 行：types 283 + extractors 897 + emitters 241 + normalizedRouting 438
- dispatcher(1780 行)/batch/hook 留主文件，二期再切 dispatch/ 四族
- 锁面 re-export：AppServerEventHandlers/DispatchOptions/coalesceAppServerEventBatch 等
- baseline 重生成 + 新文件登记豁免
验证：tsc 零 error；13 测试文件 144 passed，4 失败与基线逐一同名（compaction/routing 既有）；gate 绿
