+++ src/features/app/hooks/useAppServerEvents.ts
@@ -1,1906 +1,44 @@
-import { useCallback, useEffect, useRef } from "react";
-import type { MutableRefObject } from "react";
-import type {
-  AppServerEvent,
-  ApprovalRequest,
-  CollaborationModeBlockedRequest,
-  CollaborationModeResolvedRequest,
-  RequestUserInputRequest,
-} from "../../../types";
-import {
-  getAppServerEventBackpressureForTests,
-  resetAppServerEventBackpressureForTests,
-  subscribeRawAppServerEvents,
-} from "../../../services/events";
-import type {
-  ConversationEngine,
-  NormalizedThreadEvent,
-} from "../../threads/contracts/conversationCurtainContracts";
-import {
-  getRealtimeAdapterByEngine,
-  inferRealtimeAdapterEngine,
-} from "../../threads/adapters/realtimeAdapterRegistry";
-import { resolveConversationAssemblyMigrationGate } from "../../threads/assembly/conversationMigrationGates";
+import type { AppServerEvent } from "../../../types";
+import { buildAgentCanvasThreadId } from "../../multi-agent/runtime/agentCanvasThread";
+import { classifyCodexEventRisk, resolveCodexEventOwnership } from "./codexEventOwnership";
+import { getAppServerEventBackpressureForTests, resetAppServerEventBackpressureForTests, subscribeRawAppServerEvents } from "../../../services/events";
+import { getRuntimeReceipt } from "../../threads/utils/runtimeModelReceipt";
+import { hasPendingSharedSessionBindingForEngine, rebindSharedSessionNativeThread, resolvePendingSharedSessionBindingForEngine, resolvePendingSharedSessionBindingForTarget, resolveSharedRuntimeControlOwner, resolveSharedSessionBindingByNativeThread, resolveSharedSessionBindingFromRuntimeOwner } from "../../shared-session/runtime/sharedSessionBridge";
 import { hydrateToolSnapshotWithEventParams } from "../../threads/adapters/toolSnapshotHydration";
-import { isGeneratedImageToolName } from "../../../utils/generatedImageArtifacts";
-import {
-  hasPendingSharedSessionBindingForEngine,
-  rebindSharedSessionNativeThread,
-  resolvePendingSharedSessionBindingForEngine,
-  resolvePendingSharedSessionBindingForTarget,
-  resolveSharedSessionBindingByNativeThread,
-  resolveSharedSessionBindingFromRuntimeOwner,
-  resolveSharedRuntimeControlOwner,
-} from "../../shared-session/runtime/sharedSessionBridge";
+import { isAgentAttempt, resolveAgentAttemptOwner } from "../../multi-agent/store/agentStore";
+import { isAppServerEventBatchConsumerEnabled, readStreamingScheduleTier } from "../../threads/utils/realtimePerfFlags";
 import { isSharedSessionSupportedEngine } from "../../shared-session/utils/sharedSessionEngines";
-import {
-  canonicalQoderThreadId,
-  parseQoderSessionIdentity,
-} from "../../threads/utils/qoderSessionIdentity";
 import { isSharedV2SendEnabled } from "../../shared-session/runtime/sharedV2SendFlag";
-import type { SharedSessionNativeBinding } from "../../shared-session/runtime/sharedSessionBridge";
-import { getActiveTurnTargetForAttempt } from "../../shared-session/target/targetStore";
-import {
-  extractRuntimeModelFromPayload,
-  getRuntimeReceipt,
-  rememberRuntimeReceipt,
-} from "../../threads/utils/runtimeModelReceipt";
-import { getNativeTurnTarget } from "../../threads/utils/nativeTurnTargetLedger";
-import { updateSharedSessionNativeBinding as updateSharedSessionNativeBindingService } from "../../shared-session/services/sharedSessions";
-import { noteThreadAppServerEventReceived } from "../../threads/utils/streamLatencyDiagnostics";
-import {
-  isAppServerEventBatchConsumerEnabled,
-  readStreamingScheduleTier,
-} from "../../threads/utils/realtimePerfFlags";
-import { useRenderScheduler } from "../../../hooks/useRenderScheduler";
-import { resolveDispatchSchedule } from "../../threads/utils/renderSchedulingPolicy";
-import {
-  classifyCodexEventRisk,
-  resolveCodexEventOwnership,
-} from "./codexEventOwnership";
 import { migrateThreadAgentEventTracking } from "./appServerEventAgentTracking";
-import { useAppServerEventBatchDispatch } from "./useAppServerEventBatchDispatch";
-import {
-  isAgentAttempt,
-  resolveAgentAttemptOwner,
-} from "../../multi-agent/store/agentStore";
-import {
-  buildAgentCanvasThreadId,
-  isAgentCanvasThreadId,
-  parseAgentCanvasThreadId,
-} from "../../multi-agent/runtime/agentCanvasThread";
+import { noteThreadAppServerEventReceived } from "../../threads/utils/streamLatencyDiagnostics";
 import { rememberCollabWorkerNativeThreadId } from "../../multi-agent/runtime/collabNativeHideRegistry";
+import { resolveDispatchSchedule } from "../../threads/utils/renderSchedulingPolicy";
+import { updateSharedSessionNativeBinding as updateSharedSessionNativeBindingService } from "../../shared-session/services/sharedSessions";
+import { useAppServerEventBatchDispatch } from "./useAppServerEventBatchDispatch";
+import { useCallback, useEffect, useRef } from "react";
+import { useRenderScheduler } from "../../../hooks/useRenderScheduler";
+import type { AppServerEventHandlers, DispatchAppServerEventBatchOptions, DispatchAppServerEventOptions, UseAppServerEventsOptions } from "./appServerEventTypes";
+import { DEFAULT_APP_SERVER_EVENT_BATCH_CHUNK_SIZE } from "./appServerEventTypes";
+import type { ThreadAgentCompletedItemTracker, ThreadAgentSnapshotItemTracker } from "./appServerEventExtractors";
+import { asStringArray, buildDshContextUsagePatch, coalesceAppServerEventBatch, extractAgentMessageDeltaPayload, extractCompactionSourceFlags, extractItemIdFromParams, extractReasoningDeltaFromParams, extractRuntimeEndedTurnMap, extractThreadIdFromParams, extractTurnIdFromParams, hasAgentMessageSnapshotText, hasThreadAgentCompletion, inferGeminiReasoningHintFromThreadId, inferRawMethodEngine, isCodexRawGeneratedImageEvent, isProviderContinuationBootstrapEvent, markThreadAgentCompletionSeen, markThreadAgentSnapshotSeen, parseCompactionReason, parseNullableTokenCount, resolveCodexOwnerThreadId, resolveEventEngine, resolveFinalizedSharedNativeThreadId, resolveLatestThreadAgentSnapshotItemId, resolveLegacyModelContextWindow, resolveThreadStartedProviderProfileId, shouldIgnoreAgentMessageSnapshot, shouldRebindSharedNativeThreadOnStartedEvent, withRealtimeItemEventContext } from "./appServerEventExtractors";
+import { asString, emitAssistantRuntimeReceipt, emitCommandOutputDelta, emitFileChangeOutputDelta, emitReasoningSummaryBoundary, emitReasoningSummaryDelta, emitReasoningTextDelta, emitTerminalInteraction, maybeCaptureRuntimeReceipt } from "./appServerEventEmitters";
+import { tryRouteNormalizedRealtimeEvent } from "./appServerEventNormalizedRouting";
+export type {
+  AgentDelta,
+  AppServerEventHandlers,
+  DispatchAppServerEventBatchOptions,
+  DispatchAppServerEventOptions,
+} from "./appServerEventTypes";
+export {
+  buildCoalescibleAppServerEventKey,
+  coalesceAppServerEventBatch,
+  isProviderContinuationBootstrapEvent,
+} from "./appServerEventExtractors";
 
 export {
   getAppServerEventBackpressureForTests,
   resetAppServerEventBackpressureForTests,
 };
-
-type AgentDelta = {
-  workspaceId: string;
-  threadId: string;
-  itemId: string;
-  delta: string;
-  turnId?: string | null;
-};
-
-type TurnErrorPayload = {
-  message: string;
-  willRetry: boolean;
-  suppressMessage?: boolean;
-  engine?: ConversationEngine | null;
-  executionTargetSnapshot?: SharedSessionNativeBinding["executionTargetSnapshot"];
-};
-
-type TurnStalledPayload = {
-  message: string;
-  reasonCode: string;
-  stage: string;
-  source: string;
-  startedAtMs: number | null;
-  timeoutMs: number | null;
-  engine?: ConversationEngine | null;
-};
-
-type AgentCompleted = {
-  workspaceId: string;
-  threadId: string;
-  itemId: string;
-  text: string;
-  turnId?: string | null;
-};
-
-export type AppServerEventHandlers = {
-  onNormalizedRealtimeEvent?: (event: NormalizedThreadEvent) => void;
-  onWorkspaceConnected?: (workspaceId: string) => void;
-  onThreadStarted?: (
-    workspaceId: string,
-    thread: Record<string, unknown>,
-  ) => void;
-  onThreadSessionIdUpdated?: (
-    workspaceId: string,
-    threadId: string,
-    sessionId: string,
-    engine?: "claude" | "opencode" | "codex" | "gemini" | "grok" | "kimi" | "pi" | "dsh" | "qoder" | null,
-    turnId?: string | null,
-  ) => void;
-  onBackgroundThreadAction?: (
-    workspaceId: string,
-    threadId: string,
-    action: string,
-  ) => void;
-  onApprovalRequest?: (request: ApprovalRequest) => void;
-  onRequestUserInput?: (request: RequestUserInputRequest) => void;
-  onModeBlocked?: (event: CollaborationModeBlockedRequest) => void;
-  onModeResolved?: (event: CollaborationModeResolvedRequest) => void;
-  onAgentMessageDelta?: (event: AgentDelta) => void;
-  onAgentMessageCompleted?: (event: AgentCompleted) => void;
-  onAppServerEvent?: (event: AppServerEvent) => void;
-  onTurnStarted?: (
-    workspaceId: string,
-    threadId: string,
-    turnId: string,
-  ) => void;
-  onSharedRuntimeTurnStarted?: (
-    threadId: string,
-    runtimeTurnId: string,
-  ) => void;
-  onTurnCompleted?: (
-    workspaceId: string,
-    threadId: string,
-    turnId: string,
-  ) => void;
-  /**
-   * pi 专属：一个 agent run 彻底 settle（pump `agent_settled` 生命周期标记）。
-   * pi 的 run 含多个原生 turn，完成音等「整轮结束」语义的消费者 MUST 监听
-   * 本信号而非 turn/completed（后者每原生 turn 一次，会连响）。
-   */
-  onThreadRunSettled?: (
-    workspaceId: string,
-    threadId: string,
-    turnId: string,
-  ) => void;
-  onProcessingHeartbeat?: (
-    workspaceId: string,
-    threadId: string,
-    pulse: number,
-  ) => void;
-  onContextCompacting?: (
-    workspaceId: string,
-    threadId: string,
-    payload: {
-      usagePercent: number | null;
-      thresholdPercent: number | null;
-      targetPercent: number | null;
-      auto?: boolean | null;
-      manual?: boolean | null;
-    },
-  ) => void;
-  onContextCompacted?: (
-    workspaceId: string,
-    threadId: string,
-    turnId: string,
-    payload?: {
-      auto?: boolean | null;
-      manual?: boolean | null;
-      /** pi compaction_end 透传；缺失为 null。 */
-      reason?: "threshold" | "overflow" | "manual" | null;
-      tokensBefore?: number | null;
-      estimatedTokensAfter?: number | null;
-    },
-  ) => void;
-  onContextCompactionFailed?: (
-    workspaceId: string,
-    threadId: string,
-    reason: string,
-  ) => void;
-  onRuntimeEnded?: (
-    workspaceId: string,
-    payload: {
-      reasonCode: string;
-      message: string;
-      affectedThreadIds: string[];
-      affectedTurnIds: string[];
-      pendingRequestCount: number;
-      hadActiveLease: boolean;
-      runtimeGeneration?: string;
-      runtimeProcessId?: number;
-      runtimeStartedAtMs?: number;
-    },
-  ) => void;
-  onTurnError?: (
-    workspaceId: string,
-    threadId: string,
-    turnId: string,
-    payload: TurnErrorPayload,
-  ) => void;
-  onTurnStalled?: (
-    workspaceId: string,
-    threadId: string,
-    turnId: string,
-    payload: TurnStalledPayload,
-  ) => void;
-  onTurnPlanUpdated?: (
-    workspaceId: string,
-    threadId: string,
-    turnId: string,
-    payload: { explanation: unknown; plan: unknown },
-  ) => void;
-  onItemStarted?: (
-    workspaceId: string,
-    threadId: string,
-    item: Record<string, unknown>,
-  ) => void;
-  onItemUpdated?: (
-    workspaceId: string,
-    threadId: string,
-    item: Record<string, unknown>,
-  ) => void;
-  onItemCompleted?: (
-    workspaceId: string,
-    threadId: string,
-    item: Record<string, unknown>,
-  ) => void;
-  /**
-   * PI 后台任务状态更新（pi-background-tasks 扩展）：receipt 快照（启动）与
-   * `<background-task-notification>` 终态唤醒。task 为 canonical 快照；
-   * notification 路径 toolId 为 null，按 task.id 关联。
-   */
-  onBackgroundTaskUpdated?: (
-    workspaceId: string,
-    threadId: string,
-    payload: {
-      toolId: string | null;
-      task: Record<string, unknown>;
-      source: string;
-    },
-  ) => void;
-  onReasoningSummaryDelta?: (
-    workspaceId: string,
-    threadId: string,
-    itemId: string,
-    delta: string,
-    engineHint?: "gemini" | "grok" | "kimi" | "pi" | "dsh" | "qoder" | null,
-    turnId?: string | null,
-  ) => void;
-  onReasoningSummaryBoundary?: (
-    workspaceId: string,
-    threadId: string,
-    itemId: string,
-    engineHint?: "gemini" | "grok" | "kimi" | "pi" | "dsh" | "qoder" | null,
-    turnId?: string | null,
-  ) => void;
-  onReasoningTextDelta?: (
-    workspaceId: string,
-    threadId: string,
-    itemId: string,
-    delta: string,
-    engineHint?: "gemini" | "grok" | "kimi" | "pi" | "dsh" | "qoder" | null,
-    turnId?: string | null,
-  ) => void;
-  onCommandOutputDelta?: (
-    workspaceId: string,
-    threadId: string,
-    itemId: string,
-    delta: string,
-    turnId?: string | null,
-  ) => void;
-  onTerminalInteraction?: (
-    workspaceId: string,
-    threadId: string,
-    itemId: string,
-    stdin: string,
-    turnId?: string | null,
-  ) => void;
-  onFileChangeOutputDelta?: (
-    workspaceId: string,
-    threadId: string,
-    itemId: string,
-    delta: string,
-    turnId?: string | null,
-  ) => void;
-  onTurnDiffUpdated?: (
-    workspaceId: string,
-    threadId: string,
-    diff: string,
-  ) => void;
-  onThreadTokenUsageUpdated?: (
-    workspaceId: string,
-    threadId: string,
-    tokenUsage: Record<string, unknown>,
-  ) => void;
-  onAssistantRuntimeReceipt?: (
-    workspaceId: string,
-    threadId: string,
-    runtimeReceipt: NonNullable<
-      Extract<import("../../../types").ConversationItem, { kind: "message" }>["runtimeReceipt"]
-    >,
-  ) => void;
-  onAccountRateLimitsUpdated?: (
-    workspaceId: string,
-    rateLimits: Record<string, unknown>,
-  ) => void;
-  /**
-   * 获取指定 workspace 当前活动的 Codex thread ID
-   * 仅用于低风险兼容展示路径，不能作为 lifecycle mutation owner。
-   */
-  getActiveCodexThreadId?: (workspaceId: string) => string | null;
-  /**
-   * Returns a thread only when the workspace has exactly one processing
-   * Codex conversation. This is the bounded fallback for owner-gated
-   * lifecycle events that lack explicit thread context.
-   */
-  getSingleProcessingCodexThreadId?: (workspaceId: string) => string | null;
-};
-
-type UseAppServerEventsOptions = {
-  useNormalizedRealtimeAdapters?: boolean;
-  /** false = 不订阅事件总线（测试/特殊宿主可关） */
-  enabled?: boolean;
-};
-
-export type DispatchAppServerEventOptions = {
-  useNormalizedRealtimeAdapters: boolean;
-  threadAgentDeltaSeenRef: MutableRefObject<Record<string, true>>;
-  threadAgentCompletedSeenRef: MutableRefObject<ThreadAgentCompletedItemTracker>;
-  threadAgentSnapshotSeenRef: MutableRefObject<ThreadAgentSnapshotItemTracker>;
-};
-
-type DispatchAppServerEventBatchOptions = DispatchAppServerEventOptions & {
-  chunkSize?: number;
-  onComplete?: () => void;
-};
-
-const DEFAULT_APP_SERVER_EVENT_BATCH_CHUNK_SIZE = 64;
-
-function asString(value: unknown): string {
-  return typeof value === "string" ? value : value ? String(value) : "";
-}
-
-function emitAssistantRuntimeReceipt(
-  handlers: AppServerEventHandlers,
-  workspaceId: string,
-  threadId: string,
-  incoming: Parameters<typeof rememberRuntimeReceipt>[2],
-) {
-  const receipt = rememberRuntimeReceipt(workspaceId, threadId, incoming);
-  if (!receipt) {
-    return;
-  }
-  handlers.onAssistantRuntimeReceipt?.(workspaceId, threadId, receipt);
-}
-
-function maybeCaptureRuntimeReceipt(
-  handlers: AppServerEventHandlers,
-  workspaceId: string,
-  method: string,
-  params: Record<string, unknown>,
-  sharedThreadId?: string | null,
-  options?: { skip?: boolean },
-) {
-  if (options?.skip) {
-    return;
-  }
-  const isRaw = method.endsWith("/raw");
-  const isTurnCompleted = method === "turn/completed";
-  if (!isRaw && !isTurnCompleted) {
-    return;
-  }
-  const rawType = asString(params.type).trim().toLowerCase();
-  const subtype = asString(params.subtype).trim().toLowerCase();
-  const isRuntimeModelSidecar = rawType === "runtime_model";
-  const isAssistantIdentity =
-    rawType === "assistant" ||
-    subtype === "assistant.message.model" ||
-    subtype.includes("assistant");
-  const isInitIdentity =
-    rawType === "system" ||
-    subtype === "system.init.model" ||
-    subtype.includes("init");
-  if (isRaw && !isRuntimeModelSidecar && !isAssistantIdentity && !isInitIdentity) {
-    return;
-  }
-  const threadId = sharedThreadId || extractThreadIdFromParams(params);
-  if (
-    !threadId ||
-    // 排除式门：Shared canonical attribution 之外，协作画布与 shared-pending
-    // 别名不入账；其余（含 native 各引擎与 shared 本体）同吃 source-rank 回写链。
-    threadId.startsWith("agent-canvas:") ||
-    threadId.includes("-pending-shared-")
-  ) {
-    return;
-  }
-  const result = asRecord(params.result);
-  const model = extractRuntimeModelFromPayload(params) ??
-    extractRuntimeModelFromPayload(result);
-  if (!model) {
-    return;
-  }
-  const modelSource = isTurnCompleted
-    ? "turn.completed"
-    : isRuntimeModelSidecar && isInitIdentity
-      ? "system.init.model"
-      : isAssistantIdentity || isRuntimeModelSidecar
-        ? "assistant.message.model"
-        : isInitIdentity
-          ? "system.init.model"
-          : "assistant.message.model";
-  emitAssistantRuntimeReceipt(handlers, workspaceId, threadId, {
-    model,
-    modelSource,
-  });
-}
-
-function asRecord(value: unknown): Record<string, unknown> | null {
-  return value && typeof value === "object"
-    ? (value as Record<string, unknown>)
-    : null;
-}
-
-function parseOptionalBoolean(value: unknown): boolean | null {
-  if (typeof value === "boolean") {
-    return value;
-  }
-  if (typeof value === "number") {
-    if (value === 1) {
-      return true;
-    }
-    if (value === 0) {
-      return false;
-    }
-  }
-  if (typeof value === "string") {
-    const normalized = value.trim().toLowerCase();
-    if (normalized === "true") {
-      return true;
-    }
-    if (normalized === "false") {
-      return false;
-    }
-  }
-  return null;
-}
-
-function extractCompactionSourceFlags(params: Record<string, unknown>) {
-  const auto = parseOptionalBoolean(params.auto ?? params.automatic);
-  const manual = parseOptionalBoolean(params.manual);
-  if (auto === null && manual === null) {
-    return null;
-  }
-  return { auto, manual };
-}
-
-/** pi compaction_end 的触发原因（threshold/overflow/manual）；缺失为 null。 */
-function parseCompactionReason(value: unknown): "threshold" | "overflow" | "manual" | null {
-  return value === "threshold" || value === "overflow" || value === "manual"
-    ? value
-    : null;
-}
-
-function parseNullableTokenCount(value: unknown): number | null {
-  if (typeof value === "number" && Number.isFinite(value) && value >= 0) {
-    return value;
-  }
-  if (typeof value === "string") {
-    const parsed = Number(value);
-    if (Number.isFinite(parsed) && parsed >= 0) {
-      return parsed;
-    }
-  }
-  return null;
-}
-
-function asStringArray(value: unknown): string[] {
-  if (!Array.isArray(value)) {
-    return [];
-  }
-  return value
-    .map((entry) => asString(entry).trim())
-    .filter((entry) => entry.length > 0);
-}
-
-function extractRuntimeEndedTurnMap(value: unknown): Map<string, string> {
-  const turnMap = new Map<string, string>();
-  if (!Array.isArray(value)) {
-    return turnMap;
-  }
-  value.forEach((entry) => {
-    const objectEntry =
-      entry && typeof entry === "object"
-        ? (entry as Record<string, unknown>)
-        : null;
-    if (!objectEntry) {
-      return;
-    }
-    const threadId = asString(
-      objectEntry.threadId ?? objectEntry.thread_id,
-    ).trim();
-    const turnId = asString(objectEntry.turnId ?? objectEntry.turn_id).trim();
-    if (!threadId || !turnId) {
-      return;
-    }
-    turnMap.set(threadId, turnId);
-  });
-  return turnMap;
-}
-
-function extractThreadIdFromParams(params: Record<string, unknown>): string {
-  const turn = (params.turn as Record<string, unknown> | undefined) ?? {};
-  const threadObj =
-    (params.thread as Record<string, unknown> | undefined) ?? {};
-  return asString(
-    params.threadId ??
-      params.thread_id ??
-      turn.threadId ??
-      turn.thread_id ??
-      threadObj.threadId ??
-      threadObj.thread_id ??
-      threadObj.id ??
-      "",
-  ).trim();
-}
-
-function resolveCodexOwnerThreadId(
-  handlers: AppServerEventHandlers,
-  workspaceId: string,
-  method: string,
-  params: Record<string, unknown>,
-): string {
-  const explicitThreadId = extractThreadIdFromParams(params);
-  const fallbackThreadId = explicitThreadId
-    ? null
-    : (handlers.getSingleProcessingCodexThreadId?.(workspaceId) ?? null);
-  const ownership = resolveCodexEventOwnership({
-    workspaceId,
-    risk: classifyCodexEventRisk(method),
-    explicitThreadId,
-    explicitTurnId: extractTurnIdFromParams(params),
-    ...(explicitThreadId ? { explicitSource: "payload" as const } : {}),
-    boundedFallbackThreadIds: fallbackThreadId ? [fallbackThreadId] : [],
-  });
-  return ownership.kind === "explicit" || ownership.kind === "boundedFallback"
-    ? ownership.threadId
-    : "";
-}
-
-function getAppServerEventMethod(payload: AppServerEvent): string {
-  return String(payload.message.method ?? "");
-}
-
-function getAppServerEventParams(
-  payload: AppServerEvent,
-): Record<string, unknown> {
-  return (payload.message.params as Record<string, unknown> | undefined) ?? {};
-}
-
-export function buildCoalescibleAppServerEventKey(
-  payload: AppServerEvent,
-): string | null {
-  const method = getAppServerEventMethod(payload);
-  const params = getAppServerEventParams(payload);
-  switch (method) {
-    case "processing/heartbeat":
-    case "thread/tokenUsage/updated":
-    case "thread/compacting":
-    case "turn/diff/updated": {
-      const threadId = extractThreadIdFromParams(params);
-      return threadId
-        ? `${payload.workspace_id}\0${method}\0${threadId}`
-        : null;
-    }
-    case "account/rateLimits/updated":
-      return `${payload.workspace_id}\0${method}`;
-    default:
-      return null;
-  }
-}
-
-export function coalesceAppServerEventBatch(
-  batch: readonly AppServerEvent[],
-): AppServerEvent[] {
-  const coalesced: AppServerEvent[] = [];
-  let previousCoalesceKey: string | null = null;
-  for (const payload of batch) {
-    const coalesceKey = buildCoalescibleAppServerEventKey(payload);
-    if (
-      coalesceKey &&
-      previousCoalesceKey === coalesceKey &&
-      coalesced.length > 0
-    ) {
-      coalesced[coalesced.length - 1] = payload;
-    } else {
-      coalesced.push(payload);
-    }
-    previousCoalesceKey = coalesceKey;
-  }
-  return coalesced;
-}
-
-function extractTurnIdFromParams(params: Record<string, unknown>): string {
-  const turn = (params.turn as Record<string, unknown> | undefined) ?? {};
-  return asString(
-    params.turnId ??
-      params.turn_id ??
-      turn.id ??
-      turn.turnId ??
-      turn.turn_id ??
-      "",
-  ).trim();
-}
-
-const PROVIDER_CONTINUATION_BOOTSTRAP_TURN_PREFIX =
-  "provider-continuation-";
-
-export function isProviderContinuationBootstrapEvent(
-  payload: AppServerEvent,
-): boolean {
-  const params = getAppServerEventParams(payload);
-  return extractTurnIdFromParams(params).startsWith(
-    PROVIDER_CONTINUATION_BOOTSTRAP_TURN_PREFIX,
-  );
-}
-
-function extractItemIdFromParams(params: Record<string, unknown>): string {
-  const turn = (params.turn as Record<string, unknown> | undefined) ?? {};
-  const itemObj = (params.item as Record<string, unknown> | undefined) ?? {};
-  const messageObj =
-    (params.message as Record<string, unknown> | undefined) ?? {};
-  const partObj = (params.part as Record<string, unknown> | undefined) ?? {};
-  const contentObj =
-    (params.content as Record<string, unknown> | undefined) ?? {};
-  return asString(
-    params.itemId ??
-      params.item_id ??
-      partObj.itemId ??
-      partObj.item_id ??
-      itemObj.id ??
-      itemObj.itemId ??
-      itemObj.item_id ??
-      messageObj.id ??
-      contentObj.itemId ??
-      contentObj.item_id ??
-      turn.itemId ??
-      turn.item_id ??
-      "",
-  ).trim();
-}
-
-function extractReasoningDeltaFromParams(
-  params: Record<string, unknown>,
-): string {
-  const partObj = (params.part as Record<string, unknown> | undefined) ?? {};
-  const itemObj = (params.item as Record<string, unknown> | undefined) ?? {};
-  const contentObj =
-    (params.content as Record<string, unknown> | undefined) ?? {};
-  return asString(
-    params.delta ??
-      params.text ??
-      params.summary ??
-      partObj.delta ??
-      partObj.text ??
-      partObj.summary ??
-      itemObj.delta ??
-      itemObj.text ??
-      itemObj.summary ??
-      contentObj.delta ??
-      contentObj.text ??
-      contentObj.summary ??
-      "",
-  ).trim();
-}
-
-function extractAgentMessageDeltaPayload(
-  method: string,
-  params: Record<string, unknown>,
-): {
-  threadId: string;
-  itemId: string;
-  delta: string;
-  turnId: string | null;
-} | null {
-  const isTextAliasMethod = method === "text:delta" || method === "text/delta";
-  const isAgentDeltaMethod =
-    method === "item/agentMessage/delta" ||
-    method === "item/agentMessage/textDelta" ||
-    method === "item/agentMessage/text/delta" ||
-    isTextAliasMethod;
-  if (!isAgentDeltaMethod) {
-    return null;
-  }
-
-  const turn = (params.turn as Record<string, unknown> | undefined) ?? {};
-  const itemObj = (params.item as Record<string, unknown> | undefined) ?? {};
-  const messageObj =
-    (params.message as Record<string, unknown> | undefined) ?? {};
-  const partObj = (params.part as Record<string, unknown> | undefined) ?? {};
-  const threadId = extractThreadIdFromParams(params);
-  const turnId = extractTurnIdFromParams(params);
-  if (
-    isTextAliasMethod &&
-    !isClaudeThreadId(threadId) &&
-    !isGeminiThreadId(threadId) &&
-    !isGrokThreadId(threadId) &&
-    !isKimiThreadId(threadId) &&
-    !isPiThreadId(threadId) &&
-    !isQoderThreadId(threadId) &&
-    !isDshThreadId(threadId)
-  ) {
-    return null;
-  }
-  const rawItemId = asString(
-    params.itemId ??
-      params.item_id ??
-      itemObj.id ??
-      messageObj.id ??
-      partObj.itemId ??
-      partObj.item_id ??
-      turn.itemId ??
-      turn.item_id ??
-      (!isTextAliasMethod ? turn.id : "") ??
-      "",
-  ).trim();
-  const itemId =
-    rawItemId || (isTextAliasMethod ? `${threadId}:text-delta` : "");
-  const delta = asString(
-    params.delta ??
-      params.text ??
-      params.output_text ??
-      params.outputText ??
-      params.content ??
-      partObj.delta ??
-      partObj.text ??
-      partObj.content ??
-      itemObj.delta ??
-      itemObj.text ??
-      itemObj.content ??
-      messageObj.delta ??
-      messageObj.text ??
-      messageObj.content ??
-      "",
-  );
-
-  if (!threadId || !itemId || !delta) {
-    return null;
-  }
-  return { threadId, itemId, delta, turnId: turnId || null };
-}
-
-function withRealtimeItemEventContext(
-  item: Record<string, unknown>,
-  params: Record<string, unknown>,
-  engineSource?: ConversationEngine,
-): Record<string, unknown> {
-  const turnId = extractTurnIdFromParams(params);
-  const existingTurnId = asString(item.turnId ?? item.turn_id).trim();
-  return {
-    ...item,
-    ...(turnId && !existingTurnId ? { turnId } : {}),
-    ...(engineSource ? { engineSource } : {}),
-  };
-}
-
-function resolveEventEngine(
-  threadId: string,
-  engineHint?: ConversationEngine | null,
-): ConversationEngine {
-  return engineHint ?? inferRealtimeAdapterEngine(threadId);
-}
-
-function cloneMessageWithThreadId(
-  message: Record<string, unknown>,
-  threadId: string,
-): Record<string, unknown> {
-  const params = (message.params as Record<string, unknown> | undefined) ?? {};
-  const nextParams: Record<string, unknown> = {
-    ...params,
-    threadId,
-    thread_id: threadId,
-  };
-  const turn = (params.turn as Record<string, unknown> | undefined) ?? null;
-  if (turn) {
-    nextParams.turn = {
-      ...turn,
-      threadId,
-      thread_id: threadId,
-    };
-  }
-  const thread = (params.thread as Record<string, unknown> | undefined) ?? null;
-  if (thread) {
-    nextParams.thread = {
-      ...thread,
-      id: threadId,
-    };
-  }
-  return {
-    ...message,
-    params: nextParams,
-  };
-}
-
-function toNumber(value: unknown): number {
-  if (typeof value === "number" && Number.isFinite(value)) {
-    return value;
-  }
-  if (typeof value === "string" && value.trim()) {
-    const parsed = Number(value);
-    if (Number.isFinite(parsed)) {
-      return parsed;
-    }
-  }
-  return 0;
-}
-
-function toOptionalNumber(value: unknown): number | null {
-  if (typeof value === "number" && Number.isFinite(value)) {
-    return value;
-  }
-  if (typeof value === "string" && value.trim()) {
-    const parsed = Number(value);
-    return Number.isFinite(parsed) ? parsed : null;
-  }
-  return null;
-}
-
-function isClaudeThreadId(threadId: string): boolean {
-  return (
-    threadId.startsWith("claude:") || threadId.startsWith("claude-pending-")
-  );
-}
-
-function resolveLegacyModelContextWindow(
-  threadId: string,
-  value: unknown,
-): number | null {
-  const parsed = toOptionalNumber(value);
-  if (parsed !== null && parsed > 0) {
-    return parsed;
-  }
-  // Claude/Codex（含 codex-pending）不伪造 200K 默认窗口：三方 provider 的真实
-  // 窗口未知（128K~1M 都有可能），伪造值只会产生误导百分比；window 未上报时
-  // 透传 null，context 指示器按「未上报」降级
-  // （fix-codex-third-party-provider-model-catalog）。其他引擎维持原行为。
-  if (
-    isClaudeThreadId(threadId) ||
-    threadId.startsWith("codex:") ||
-    threadId.startsWith("codex-pending-")
-  ) {
-    return null;
-  }
-  return 200000;
-}
-
-function isGeminiThreadId(threadId: string): boolean {
-  return (
-    threadId.startsWith("gemini:") || threadId.startsWith("gemini-pending-")
-  );
-}
-
-function isKimiThreadId(threadId: string): boolean {
-  return (
-    threadId.startsWith("kimi:") || threadId.startsWith("kimi-pending-")
-  );
-}
-
-function isPiThreadId(threadId: string): boolean {
-  return threadId.startsWith("pi:") || threadId.startsWith("pi-pending-");
-}
-
-function isQoderThreadId(threadId: string): boolean {
-  return threadId.startsWith("qoder:") || threadId.startsWith("qoder-pending-");
-}
-
-function isGrokThreadId(threadId: string): boolean {
-  return (
-    threadId.startsWith("grok:") || threadId.startsWith("grok-pending-")
-  );
-}
-
-function isDshThreadId(threadId: string): boolean {
-  return (
-    threadId.startsWith("dsh:") || threadId.startsWith("dsh-pending-")
-  );
-}
-
-function inferGeminiReasoningHintFromThreadId(
-  threadId: string,
-): "gemini" | "grok" | "kimi" | "pi" | "dsh" | "qoder" | null {
-  if (!threadId) {
-    return null;
-  }
-  if (isGrokThreadId(threadId)) {
-    return "grok";
-  }
-  if (isKimiThreadId(threadId)) {
-    return "kimi";
-  }
-  if (isPiThreadId(threadId)) {
-    return "pi";
-  }
-  if (isQoderThreadId(threadId)) {
-    return "qoder";
-  }
-  if (isDshThreadId(threadId)) {
-    return "dsh";
-  }
-  return isGeminiThreadId(threadId) ? "gemini" : null;
-}
-
-function inferRawMethodEngine(
-  method: string,
-): "claude" | "codex" | "gemini" | "grok" | "kimi" | "opencode" | "pi" | "dsh" | "qoder" | undefined {
-  switch (method) {
-    case "claude/raw":
-      return "claude";
-    case "codex/raw":
-      return "codex";
-    case "gemini/raw":
-      return "gemini";
-    case "grok/raw":
-      return "grok";
-    case "kimi/raw":
-      return "kimi";
-    case "opencode/raw":
-      return "opencode";
-    case "pi/raw":
-      return "pi";
-    case "qoder/raw":
-      return "qoder";
-    case "dsh/raw":
-      return "dsh";
-    default:
-      return undefined;
-  }
-}
-
-function isCodexRawGeneratedImageEvent(
-  method: string,
-  params: Record<string, unknown>,
-): boolean {
-  if (method !== "codex/raw") {
-    return false;
-  }
-  const rawEntryType = asString(params.type ?? "")
-    .trim()
-    .toLowerCase();
-  if (rawEntryType !== "event_msg" && rawEntryType !== "response_item") {
-    return false;
-  }
-  const payload =
-    params.payload && typeof params.payload === "object"
-      ? (params.payload as Record<string, unknown>)
-      : null;
-  if (!payload) {
-    return false;
-  }
-  const payloadType = asString(payload.type ?? "")
-    .trim()
-    .toLowerCase();
-  if (payloadType === "function_call") {
-    return isGeneratedImageToolName(
-      asString(payload.name ?? payload.tool ?? ""),
-    );
-  }
-  return (
-    payloadType === "image_generation_call" ||
-    payloadType === "image_generation_end" ||
-    payloadType === "function_call_output"
-  );
-}
-
-function shouldRebindSharedNativeThreadOnStartedEvent(
-  engine: "claude" | "opencode" | "codex" | "gemini" | "grok" | "kimi" | "pi" | "qoder",
-): boolean {
-  // Claude 与 local CLIs 在 thread/started 上可能从 pending 占位收敛到
-  // `engine:{sessionId}`；Codex 使用 raw thread id，不在此路径做前缀 rebind。
-  // Qoder 终态 id 额外带 distribution：`qoder:<profile>:<sessionId>`。
-  return (
-    engine === "claude" ||
-    engine === "kimi" ||
-    engine === "grok" ||
-    engine === "pi" ||
-    engine === "opencode" ||
-    engine === "qoder"
-  );
-}
-
-function readEventProviderProfileId(
-  params: Record<string, unknown>,
-  thread: Record<string, unknown> | null,
-): string | null {
-  const owner =
-    params.sharedOwner && typeof params.sharedOwner === "object"
-      ? (params.sharedOwner as Record<string, unknown>)
-      : null;
-  const snapshot =
-    owner?.executionTargetSnapshot &&
-    typeof owner.executionTargetSnapshot === "object"
-      ? (owner.executionTargetSnapshot as Record<string, unknown>)
-      : null;
-  const candidates = [
-    params.providerProfileId,
-    params.provider_profile_id,
-    thread?.providerProfileId,
-    thread?.provider_profile_id,
-    owner?.providerProfileId,
-    owner?.provider_profile_id,
-    snapshot?.providerProfileId,
-    snapshot?.provider_profile_id,
-  ];
-  for (const value of candidates) {
-    if (typeof value === "string" && value.trim()) {
-      return value.trim();
-    }
-  }
-  return null;
-}
-
-function resolveThreadStartedProviderProfileId(params: {
-  params: Record<string, unknown>;
-  thread: Record<string, unknown> | null;
-  threadId: string;
-  sessionId: string;
-  eventEngine: string | null;
-}): string | null {
-  const fromEvent = readEventProviderProfileId(params.params, params.thread);
-  if (fromEvent) {
-    return fromEvent;
-  }
-  if (params.eventEngine !== "qoder") {
-    return null;
-  }
-  for (const value of [params.threadId, params.sessionId]) {
-    const identity = parseQoderSessionIdentity(value);
-    if (identity && !identity.isLegacy) {
-      return identity.providerProfileId;
-    }
-  }
-  return null;
-}
-
-function resolveFinalizedSharedNativeThreadId(
-  eventEngine:
-    | "claude"
-    | "opencode"
-    | "codex"
-    | "gemini"
-    | "grok"
-    | "kimi"
-    | "pi"
-    | "qoder",
-  sessionId: string,
-  providerProfileId: string | null,
-): string | null {
-  if (eventEngine === "qoder") {
-    return canonicalQoderThreadId(sessionId, providerProfileId);
-  }
-  return `${eventEngine}:${sessionId}`;
-}
-
-function isAgentMessageSnapshotMethod(method: string): boolean {
-  return method === "item/started" || method === "item/updated";
-}
-
-function shouldIgnoreAgentMessageSnapshot(params: {
-  threadId: string;
-  itemType: string;
-  method: string;
-  threadAgentDeltaSeenRef: MutableRefObject<Record<string, true>>;
-}): boolean {
-  const { threadId, itemType, method, threadAgentDeltaSeenRef } = params;
-  if (itemType !== "agentMessage" || !isAgentMessageSnapshotMethod(method)) {
-    return false;
-  }
-  if (isClaudeThreadId(threadId)) {
-    return method !== "item/updated";
-  }
-  return Boolean(threadAgentDeltaSeenRef.current[threadId]);
-}
-
-function hasAgentMessageSnapshotText(item: Record<string, unknown>): boolean {
-  const text = asString(
-    item.text ?? item.content ?? item.output_text ?? item.outputText ?? "",
-  ).trim();
-  return text.length > 0;
-}
-
-function optionalFiniteNumber(value: unknown): number | null {
-  if (typeof value === "number" && Number.isFinite(value)) {
-    return value;
-  }
-  if (typeof value === "string" && value.trim()) {
-    const parsed = Number(value);
-    return Number.isFinite(parsed) ? parsed : null;
-  }
-  return null;
-}
-
-function buildDshContextUsagePatch(
-  params: Record<string, unknown>,
-): Record<string, unknown> | null {
-  const pressure =
-    params.contextPressure && typeof params.contextPressure === "object"
-      ? (params.contextPressure as Record<string, unknown>)
-      : null;
-  const breakdown =
-    params.contextBreakdown && typeof params.contextBreakdown === "object"
-      ? (params.contextBreakdown as Record<string, unknown>)
-      : null;
-  if (!pressure && !breakdown) {
-    return null;
-  }
-  const patch: Record<string, unknown> = {
-    contextUsageSource: "dsh-context-pressure",
-    contextUsageFreshness: "live",
-  };
-  if (pressure) {
-    const used =
-      optionalFiniteNumber(pressure.projectedTokens) ??
-      optionalFiniteNumber(pressure.pressureTokens);
-    const window = optionalFiniteNumber(pressure.contextWindow);
-    if (used !== null) {
-      patch.contextUsedTokens = used;
-    }
-    if (window !== null && window > 0) {
-      patch.modelContextWindow = window;
-    }
-    if (used !== null && window !== null && window > 0) {
-      const percent = (used / window) * 100;
-      patch.contextUsedPercent = percent;
-      patch.contextRemainingPercent = Math.max(100 - percent, 0);
-    }
-  }
-  if (breakdown) {
-    const rows = [
-      ["system", breakdown.systemTokens],
-      ["tools", breakdown.toolsTokens],
-      ["messages", breakdown.messageTokens],
-    ]
-      .map(([name, tokens]) => {
-        const value = optionalFiniteNumber(tokens);
-        return value === null ? null : { name, tokens: value };
-      })
-      .filter((row): row is { name: string; tokens: number } => row !== null);
-    if (rows.length > 0) {
-      patch.contextCategoryUsages = rows;
-    }
-  }
-  return Object.keys(patch).length > 2 ? patch : patch;
-}
-
-function extractTokenUsageFromNormalizedEvent(
-  event: NormalizedThreadEvent,
-): Record<string, unknown> | null {
-  const usageFromItem =
-    event.rawItem &&
-    typeof event.rawItem.usage === "object" &&
-    event.rawItem.usage
-      ? (event.rawItem.usage as Record<string, unknown>)
-      : null;
-  const usage = event.rawUsage ?? usageFromItem;
-  if (!usage) {
-    return null;
-  }
-
-  const inputTokens = toNumber(usage.input_tokens ?? usage.inputTokens);
-  const outputTokens = toNumber(usage.output_tokens ?? usage.outputTokens);
-  const cachedInputTokens = toNumber(
-    usage.cached_input_tokens ??
-      usage.cache_read_input_tokens ??
-      usage.cachedInputTokens ??
-      usage.cacheReadInputTokens,
-  );
-  const modelContextWindow = toNumber(
-    usage.model_context_window ?? usage.modelContextWindow,
-  );
-  if (inputTokens <= 0 && outputTokens <= 0 && cachedInputTokens <= 0) {
-    return null;
-  }
-  const contextUsedPercent = toOptionalNumber(
-    usage.context_used_percent ?? usage.contextUsedPercent,
-  );
-  const contextRemainingPercent = toOptionalNumber(
-    usage.context_remaining_percent ?? usage.contextRemainingPercent,
-  );
-  const contextUsedTokens = toOptionalNumber(
-    usage.context_used_tokens ?? usage.contextUsedTokens,
-  );
-  const contextUsageSource =
-    typeof (usage.context_usage_source ?? usage.contextUsageSource) === "string"
-      ? String(usage.context_usage_source ?? usage.contextUsageSource)
-      : null;
-  const contextUsageFreshness =
-    typeof (usage.context_usage_freshness ?? usage.contextUsageFreshness) ===
-    "string"
-      ? String(usage.context_usage_freshness ?? usage.contextUsageFreshness)
-      : null;
-  return {
-    total: {
-      inputTokens,
-      outputTokens,
-      cachedInputTokens,
-      totalTokens: inputTokens + outputTokens,
-    },
-    last: {
-      inputTokens,
-      outputTokens,
-      cachedInputTokens,
-      totalTokens: inputTokens + outputTokens,
-    },
-    modelContextWindow: modelContextWindow > 0 ? modelContextWindow : null,
-    contextUsageSource,
-    contextUsageFreshness,
-    contextUsedTokens:
-      contextUsedTokens !== null && contextUsedTokens >= 0
-        ? contextUsedTokens
-        : null,
-    contextUsedPercent:
-      contextUsedPercent !== null && contextUsedPercent >= 0
-        ? contextUsedPercent
-        : null,
-    contextRemainingPercent:
-      contextRemainingPercent !== null && contextRemainingPercent >= 0
-        ? contextRemainingPercent
-        : null,
-  };
-}
-
-type ThreadAgentCompletedItemTracker = Record<string, Record<string, true>>;
-type ThreadAgentSnapshotItemTracker = Record<string, Record<string, true>>;
-
-function resolveAgentCompletionKey(itemId: string, text: string): string {
-  const normalizedItemId = itemId.trim();
-  if (normalizedItemId) {
-    return `item:${normalizedItemId}`;
-  }
-  const normalizedText = text.trim();
-  if (normalizedText) {
-    return `text:${normalizedText}`;
-  }
-  return "";
-}
-
-function hasThreadAgentCompletion(
-  trackerRef: MutableRefObject<ThreadAgentCompletedItemTracker>,
-  threadId: string,
-): boolean {
-  const threadTracker = trackerRef.current[threadId];
-  return Boolean(threadTracker && Object.keys(threadTracker).length > 0);
-}
-
-function markThreadAgentCompletionSeen(
-  trackerRef: MutableRefObject<ThreadAgentCompletedItemTracker>,
-  threadId: string,
-  itemId: string,
-  text: string,
-): boolean {
-  const completionKey = resolveAgentCompletionKey(itemId, text);
-  if (!completionKey) {
-    return true;
-  }
-  const threadTracker = trackerRef.current[threadId] ?? {};
-  if (threadTracker[completionKey]) {
-    return false;
-  }
-  threadTracker[completionKey] = true;
-  trackerRef.current[threadId] = threadTracker;
-  return true;
-}
-
-function markThreadAgentSnapshotSeen(
-  trackerRef: MutableRefObject<ThreadAgentSnapshotItemTracker>,
-  threadId: string,
-  itemId: string,
-): void {
-  if (!threadId || !itemId) {
-    return;
-  }
-  const threadTracker = trackerRef.current[threadId] ?? {};
-  threadTracker[itemId] = true;
-  trackerRef.current[threadId] = threadTracker;
-}
-
-function hasThreadAgentSnapshotSeen(
-  trackerRef: MutableRefObject<ThreadAgentSnapshotItemTracker>,
-  threadId: string,
-  itemId: string,
-): boolean {
-  if (!threadId || !itemId) {
-    return false;
-  }
-  return Boolean(trackerRef.current[threadId]?.[itemId]);
-}
-
-function resolveLatestThreadAgentSnapshotItemId(
-  trackerRef: MutableRefObject<ThreadAgentSnapshotItemTracker>,
-  threadId: string,
-): string | null {
-  const itemIds = Object.keys(trackerRef.current[threadId] ?? {});
-  return itemIds[itemIds.length - 1] ?? null;
-}
-
-function emitReasoningSummaryDelta(
-  handlers: AppServerEventHandlers,
-  workspaceId: string,
-  threadId: string,
-  itemId: string,
-  delta: string,
-  engineHint: "gemini" | "grok" | "kimi" | "pi" | "dsh" | "qoder" | null,
-  turnId: string | null,
-): void {
-  if (turnId) {
-    handlers.onReasoningSummaryDelta?.(
-      workspaceId,
-      threadId,
-      itemId,
-      delta,
-      engineHint,
-      turnId,
-    );
-    return;
-  }
-  if (engineHint) {
-    handlers.onReasoningSummaryDelta?.(
-      workspaceId,
-      threadId,
-      itemId,
-      delta,
-      engineHint,
-    );
-    return;
-  }
-  handlers.onReasoningSummaryDelta?.(workspaceId, threadId, itemId, delta);
-}
-
-function emitReasoningSummaryBoundary(
-  handlers: AppServerEventHandlers,
-  workspaceId: string,
-  threadId: string,
-  itemId: string,
-  engineHint: "gemini" | "grok" | "kimi" | "pi" | "dsh" | "qoder" | null,
-  turnId: string | null,
-): void {
-  if (turnId) {
-    handlers.onReasoningSummaryBoundary?.(
-      workspaceId,
-      threadId,
-      itemId,
-      engineHint,
-      turnId,
-    );
-    return;
-  }
-  if (engineHint) {
-    handlers.onReasoningSummaryBoundary?.(
-      workspaceId,
-      threadId,
-      itemId,
-      engineHint,
-    );
-    return;
-  }
-  handlers.onReasoningSummaryBoundary?.(workspaceId, threadId, itemId);
-}
-
-function emitReasoningTextDelta(
-  handlers: AppServerEventHandlers,
-  workspaceId: string,
-  threadId: string,
-  itemId: string,
-  delta: string,
-  engineHint: "gemini" | "grok" | "kimi" | "pi" | "dsh" | "qoder" | null,
-  turnId: string | null,
-): void {
-  if (turnId) {
-    handlers.onReasoningTextDelta?.(
-      workspaceId,
-      threadId,
-      itemId,
-      delta,
-      engineHint,
-      turnId,
-    );
-    return;
-  }
-  if (engineHint) {
-    handlers.onReasoningTextDelta?.(
-      workspaceId,
-      threadId,
-      itemId,
-      delta,
-      engineHint,
-    );
-    return;
-  }
-  handlers.onReasoningTextDelta?.(workspaceId, threadId, itemId, delta);
-}
-
-function emitCommandOutputDelta(
-  handlers: AppServerEventHandlers,
-  workspaceId: string,
-  threadId: string,
-  itemId: string,
-  delta: string,
-  turnId: string | null,
-): void {
-  if (turnId) {
-    handlers.onCommandOutputDelta?.(
-      workspaceId,
-      threadId,
-      itemId,
-      delta,
-      turnId,
-    );
-    return;
-  }
-  handlers.onCommandOutputDelta?.(workspaceId, threadId, itemId, delta);
-}
-
-function emitFileChangeOutputDelta(
-  handlers: AppServerEventHandlers,
-  workspaceId: string,
-  threadId: string,
-  itemId: string,
-  delta: string,
-  turnId: string | null,
-): void {
-  if (turnId) {
-    handlers.onFileChangeOutputDelta?.(
-      workspaceId,
-      threadId,
-      itemId,
-      delta,
-      turnId,
-    );
-    return;
-  }
-  handlers.onFileChangeOutputDelta?.(workspaceId, threadId, itemId, delta);
-}
-
-function emitTerminalInteraction(
-  handlers: AppServerEventHandlers,
-  workspaceId: string,
-  threadId: string,
-  itemId: string,
-  stdin: string,
-  turnId: string | null,
-): void {
-  if (turnId) {
-    handlers.onTerminalInteraction?.(
-      workspaceId,
-      threadId,
-      itemId,
-      stdin,
-      turnId,
-    );
-    return;
-  }
-  handlers.onTerminalInteraction?.(workspaceId, threadId, itemId, stdin);
-}
-
-function routeNormalizedRealtimeEvent({
-  handlers,
-  workspaceId,
-  event,
-  threadAgentDeltaSeenRef,
-  threadAgentCompletedSeenRef,
-  threadAgentSnapshotSeenRef,
-}: {
-  handlers: AppServerEventHandlers;
-  workspaceId: string;
-  event: NormalizedThreadEvent;
-  threadAgentDeltaSeenRef: MutableRefObject<Record<string, true>>;
-  threadAgentCompletedSeenRef: MutableRefObject<ThreadAgentCompletedItemTracker>;
-  threadAgentSnapshotSeenRef: MutableRefObject<ThreadAgentSnapshotItemTracker>;
-}): boolean {
-  const threadId = event.threadId;
-  const itemId = event.item.id;
-  const turnId = event.turnId ?? null;
-  const shouldRouteDirectly =
-    Boolean(handlers.onNormalizedRealtimeEvent) &&
-    (event.engine === "codex" ||
-      event.threadId.startsWith("shared:") ||
-      // 协作节点 Inspector 幕布：与 shared 同源 normalized 路由
-      event.threadId.startsWith("agent-canvas:"));
-  switch (event.operation) {
-    case "itemStarted":
-      if (
-        event.engine === "codex" &&
-        event.item.kind === "message" &&
-        event.item.role === "assistant"
-      ) {
-        markThreadAgentSnapshotSeen(
-          threadAgentSnapshotSeenRef,
-          threadId,
-          itemId,
-        );
-      }
-      if (shouldRouteDirectly) {
-        handlers.onNormalizedRealtimeEvent?.(event);
-        return true;
-      }
-      if (event.rawItem) {
-        handlers.onItemStarted?.(workspaceId, threadId, event.rawItem);
-        return true;
-      }
-      return false;
-    case "itemUpdated":
-      if (
-        event.engine === "codex" &&
-        event.item.kind === "message" &&
-        event.item.role === "assistant"
-      ) {
-        markThreadAgentSnapshotSeen(
-          threadAgentSnapshotSeenRef,
-          threadId,
-          itemId,
-        );
-      }
-      if (shouldRouteDirectly) {
-        handlers.onNormalizedRealtimeEvent?.(event);
-        return true;
-      }
-      if (event.rawItem) {
-        handlers.onItemUpdated?.(workspaceId, threadId, event.rawItem);
-        return true;
-      }
-      return false;
-    case "itemCompleted":
-      if (shouldRouteDirectly) {
-        handlers.onNormalizedRealtimeEvent?.(event);
-        const tokenUsage = extractTokenUsageFromNormalizedEvent(event);
-        if (tokenUsage) {
-          handlers.onThreadTokenUsageUpdated?.(
-            workspaceId,
-            threadId,
-            tokenUsage,
-          );
-        }
-        return true;
-      }
-      if (event.rawItem) {
-        handlers.onItemCompleted?.(workspaceId, threadId, event.rawItem);
-        const tokenUsage = extractTokenUsageFromNormalizedEvent(event);
-        if (tokenUsage) {
-          handlers.onThreadTokenUsageUpdated?.(
-            workspaceId,
-            threadId,
-            tokenUsage,
-          );
-        }
-        return true;
-      }
-      return false;
-    case "appendAgentMessageDelta": {
-      if (
-        shouldIgnoreAgentMessageSnapshot({
-          threadId,
-          itemType: "agentMessage",
-          method: event.sourceMethod,
-          threadAgentDeltaSeenRef,
-        })
-      ) {
-        // Claude should accept growing item/updated snapshots so the curtain can
-        // reveal long Markdown before completion, but item/started snapshots are
-        // still treated as setup noise. Other engines only ignore snapshot aliases
-        // after a real streaming delta has already arrived.
-        return true;
-      }
-      const delta =
-        event.delta ?? (event.item.kind === "message" ? event.item.text : "");
-      if (!delta) {
-        return false;
-      }
-      if (
-        event.engine === "codex" &&
-        hasThreadAgentSnapshotSeen(threadAgentSnapshotSeenRef, threadId, itemId)
-      ) {
-        return true;
-      }
-      markThreadAgentSnapshotSeen(
-        threadAgentSnapshotSeenRef,
-        threadId,
-        itemId,
-      );
-      threadAgentDeltaSeenRef.current[threadId] = true;
-      if (shouldRouteDirectly) {
-        handlers.onNormalizedRealtimeEvent?.({
-          ...event,
-          delta,
-          item:
-            event.item.kind === "message"
-              ? { ...event.item, text: delta }
-              : event.item,
-        });
-        return true;
-      }
-      handlers.onAgentMessageDelta?.({
-        workspaceId,
-        threadId,
-        itemId,
-        delta,
-        ...(turnId ? { turnId } : {}),
-      });
-      return true;
-    }
-    case "completeAgentMessage": {
-      const text = event.item.kind === "message" ? event.item.text : "";
-      const tokenUsage = extractTokenUsageFromNormalizedEvent(event);
-      if (tokenUsage) {
-        handlers.onThreadTokenUsageUpdated?.(workspaceId, threadId, tokenUsage);
-      }
-      if (
-        !markThreadAgentCompletionSeen(
-          threadAgentCompletedSeenRef,
-          threadId,
-          itemId,
-          text,
-        )
-      ) {
-        return true;
-      }
-      if (shouldRouteDirectly) {
-        handlers.onNormalizedRealtimeEvent?.(event);
-        return true;
-      }
-      if (event.rawItem) {
-        handlers.onItemCompleted?.(workspaceId, threadId, event.rawItem);
-      }
-      handlers.onAgentMessageCompleted?.({
-        workspaceId,
-        threadId,
-        itemId,
-        text,
-        ...(turnId ? { turnId } : {}),
-      });
-      return true;
-    }
-    case "appendReasoningSummaryDelta": {
-      const delta = event.delta ?? "";
-      if (!delta) {
-        return false;
-      }
-      if (shouldRouteDirectly) {
-        handlers.onNormalizedRealtimeEvent?.({
-          ...event,
-          delta,
-          item:
-            event.item.kind === "reasoning"
-              ? {
-                  ...event.item,
-                  summary: delta,
-                }
-              : event.item,
-        });
-        return true;
-      }
-      emitReasoningSummaryDelta(
-        handlers,
-        workspaceId,
-        threadId,
-        itemId,
-        delta,
-        event.engine === "gemini" || event.engine === "grok" || event.engine === "kimi" || event.engine === "pi" || event.engine === "qoder" ? event.engine : null,
-        turnId,
-      );
-      return true;
-    }
-    case "appendReasoningSummaryBoundary":
-      if (shouldRouteDirectly) {
-        handlers.onNormalizedRealtimeEvent?.(event);
-        return true;
-      }
-      emitReasoningSummaryBoundary(
-        handlers,
-        workspaceId,
-        threadId,
-        itemId,
-        event.engine === "gemini" || event.engine === "grok" || event.engine === "kimi" || event.engine === "pi" || event.engine === "qoder" ? event.engine : null,
-        turnId,
-      );
-      return true;
-    case "appendReasoningContentDelta": {
-      const delta = event.delta ?? "";
-      if (!delta) {
-        return false;
-      }
-      if (shouldRouteDirectly) {
-        handlers.onNormalizedRealtimeEvent?.({
-          ...event,
-          delta,
-          item:
-            event.item.kind === "reasoning"
-              ? {
-                  ...event.item,
-                  content: delta,
-                }
-              : event.item,
-        });
-        return true;
-      }
-      emitReasoningTextDelta(
-        handlers,
-        workspaceId,
-        threadId,
-        itemId,
-        delta,
-        event.engine === "gemini" || event.engine === "grok" || event.engine === "kimi" || event.engine === "pi" || event.engine === "qoder" ? event.engine : null,
-        turnId,
-      );
-      return true;
-    }
-    case "appendToolOutputDelta": {
-      const delta = event.delta ?? "";
-      if (!delta || event.item.kind !== "tool") {
-        return false;
-      }
-      if (shouldRouteDirectly) {
-        handlers.onNormalizedRealtimeEvent?.({
-          ...event,
-          delta,
-          item: {
-            ...event.item,
-            output: delta,
-          },
-        });
-        return true;
-      }
-      if (event.item.toolType === "fileChange") {
-        emitFileChangeOutputDelta(
-          handlers,
-          workspaceId,
-          threadId,
-          itemId,
-          delta,
-          turnId,
-        );
-      } else {
-        emitCommandOutputDelta(
-          handlers,
-          workspaceId,
-          threadId,
-          itemId,
-          delta,
-          turnId,
-        );
-      }
-      return true;
-    }
-    default:
-      return false;
-  }
-}
-
-function tryRouteNormalizedRealtimeEvent({
-  handlers,
-  workspaceId,
-  message,
-  engineOverride,
-  threadIdOverride,
-  sharedBinding,
-  threadAgentDeltaSeenRef,
-  threadAgentCompletedSeenRef,
-  threadAgentSnapshotSeenRef,
-}: {
-  handlers: AppServerEventHandlers;
-  workspaceId: string;
-  message: Record<string, unknown>;
-  engineOverride?: "claude" | "codex" | "gemini" | "grok" | "kimi" | "opencode" | "pi" | "dsh" | "qoder";
-  threadIdOverride?: string;
-  sharedBinding?: SharedSessionNativeBinding | null;
-  threadAgentDeltaSeenRef: MutableRefObject<Record<string, true>>;
-  threadAgentCompletedSeenRef: MutableRefObject<ThreadAgentCompletedItemTracker>;
-  threadAgentSnapshotSeenRef: MutableRefObject<ThreadAgentSnapshotItemTracker>;
-}): boolean {
-  const params = (message.params as Record<string, unknown> | undefined) ?? {};
-  const turn = (params.turn as Record<string, unknown> | undefined) ?? {};
-  const rawThreadId = asString(
-    params.threadId ??
-      params.thread_id ??
-      turn.threadId ??
-      turn.thread_id ??
-      "",
-  );
-  const effectiveThreadId = threadIdOverride || rawThreadId;
-  if (!effectiveThreadId) {
-    return false;
-  }
-  const engine =
-    engineOverride ?? inferRealtimeAdapterEngine(effectiveThreadId);
-  const migrationGate = resolveConversationAssemblyMigrationGate(engine);
-  if (migrationGate && !migrationGate.assemblerEnabled) {
-    return false;
-  }
-  const adapter = getRealtimeAdapterByEngine(engine);
-  const shouldInjectThreadId = Boolean(threadIdOverride);
-  const normalized = adapter.mapEvent({
-    workspaceId,
-    message: shouldInjectThreadId
-      ? cloneMessageWithThreadId(message, effectiveThreadId)
-      : message,
-  });
-  if (!normalized) {
-    return false;
-  }
-  const isSharedOwnerProjection = effectiveThreadId.startsWith("shared:");
-  const isAgentCanvasProjection = isAgentCanvasThreadId(effectiveThreadId);
-  // Shared 投影读同一 store；native 打开同款注入，Ⓡ 尾巴与展开面板随之可用。
-  const runtimeReceipt =
-    isSharedOwnerProjection || (!isAgentCanvasProjection && !sharedBinding)
-      ? getRuntimeReceipt(workspaceId, effectiveThreadId)
-      : null;
-  // Native turn-target：codex 等走 normalized 直达路由的引擎也按发送边界账本
-  // 标注本轮 provenance（shared canonical / attempt 注入优先级不变）。
-  const nativeExecutionTargetSnapshot =
-    !isSharedOwnerProjection && !isAgentCanvasProjection
-      ? getNativeTurnTarget(workspaceId, effectiveThreadId)
-      : null;
-  if (
-    shouldInjectThreadId ||
-    isSharedOwnerProjection ||
-    Boolean(nativeExecutionTargetSnapshot)
-  ) {
-    // agent-canvas: 事件写到隔离 thread，但 activeTurn 挂在 shared: 上
-    const activeTurnThreadId = isAgentCanvasProjection
-      ? parseAgentCanvasThreadId(effectiveThreadId)?.sharedThreadId ??
-        effectiveThreadId
-      : effectiveThreadId;
-    const executionTargetSnapshot =
-      sharedBinding?.executionTargetSnapshot ??
-      (sharedBinding?.attemptId
-        ? getActiveTurnTargetForAttempt(
-            workspaceId,
-            activeTurnThreadId,
-            sharedBinding.attemptId,
-          )
-        : null) ?? nativeExecutionTargetSnapshot;
-    if (shouldInjectThreadId || isSharedOwnerProjection) {
-      normalized.threadId = effectiveThreadId;
-    }
-    normalized.item =
-      // context-event 留痕不携带 engineSource（合成 item，非引擎产物）。
-      normalized.item.kind === "context-event"
-        ? normalized.item
-        : {
-            ...normalized.item,
-            engineSource: engine,
-            ...(normalized.item.kind === "message" &&
-            normalized.item.role === "assistant"
-              ? {
-                  ...(executionTargetSnapshot ? { executionTargetSnapshot } : {}),
-                  ...(runtimeReceipt ? { runtimeReceipt } : {}),
-                }
-              : {}),
-          };
-    if (normalized.rawItem) {
-      normalized.rawItem = {
-        ...normalized.rawItem,
-        engineSource: engine,
-      };
-    }
-  }
-  return routeNormalizedRealtimeEvent({
-    handlers,
-    workspaceId,
-    event: normalized,
-    threadAgentDeltaSeenRef,
-    threadAgentCompletedSeenRef,
-    threadAgentSnapshotSeenRef,
-  });
-}
-
-/**
- * Module-level dispatcher for a single `AppServerEvent`.
- *
- * Extracted from the `useAppServerEvents` `useEffect` callback so both the
- * fallback raw subscription and the v2 scheduled consumer share one routing
- * path. Sharing matters: a per-event closure copy would double-register
- * handlers and cause duplicate reducer dispatches when multiple channels are
- * active.
- *
- * All closure-captured state (handlers, refs, runtime options) is
- * passed explicitly so the dispatcher has no hidden dependencies and is
- * unit-testable in isolation.
- */
 export function dispatchAppServerEvent(
   handlers: AppServerEventHandlers,
   payload: AppServerEvent,
